Refactor repositories
