+++ lib/repositories/analytics-repository.ts
@@ -1,26 +1,9 @@
-import {
-  DescribeStatementCommand,
-  ExecuteStatementCommand,
-  RedshiftDataClient,
-  StatusString,
-} from '@aws-sdk/client-redshift-data';
+import { RedshiftDataClient } from '@aws-sdk/client-redshift-data';
 import Logger from 'bunyan';
 
-import { sleep } from '../util/time';
-import { BaseAnalyticsRepository, ExecutionConfigs } from './base';
+import { BaseRedshiftRepository, SharedConfigs, TimestampThreshold } from './base';
 
-export enum TimestampThreshold {
-  ONE_MONTH = "'1 MONTH'",
-  TWO_MONTHS = "'2 MONTHS'",
-}
-
-export type SharedConfigs = {
-  Database: string;
-  ClusterIdentifier: string;
-  SecretArn: string;
-};
-
-export class AnalyticsRepository implements BaseAnalyticsRepository {
+export class AnalyticsRepository extends BaseRedshiftRepository {
   static log: Logger;
 
   static create(configs: SharedConfigs): AnalyticsRepository {
@@ -32,7 +15,9 @@ export class AnalyticsRepository implements BaseAnalyticsRepository {
     return new AnalyticsRepository(new RedshiftDataClient({}), configs);
   }
 
-  constructor(readonly client: RedshiftDataClient, private readonly configs: SharedConfigs) {}
+  constructor(readonly client: RedshiftDataClient, private readonly _configs: SharedConfigs) {
+    super(client, _configs);
+  }
 
   public async cleanUpTable(
     tableName: string,
@@ -46,36 +31,7 @@ export class AnalyticsRepository implements BaseAnalyticsRepository {
     // immediately reclaim storage space
     const vacuumSql = `VACUUM DELETE ONLY ${tableName}`;
 
-    await this.executeStatement(deleteSql, { waitTimeMs: 10_000 });
-    await this.executeStatement(vacuumSql, { waitTimeMs: 2_000 });
-  }
-
-  private async executeStatement(sql: string, executionConfigs?: ExecutionConfigs): Promise<void> {
-    const response = await this.client.send(new ExecuteStatementCommand({ ...this.configs, Sql: sql }));
-    const stmtId = response.Id;
-
-    for (;;) {
-      const status = await this.client.send(new DescribeStatementCommand({ Id: stmtId }));
-      if (status.Error) {
-        AnalyticsRepository.log.error({ error: status.Error }, 'Failed to execute command');
-        throw new Error(status.Error);
-      }
-      if (status.Status === StatusString.ABORTED || status.Status === StatusString.FAILED) {
-        AnalyticsRepository.log.error({ error: status.Error }, 'Failed to execute command');
-        throw new Error(status.Error);
-      } else if (
-        status.Status === StatusString.PICKED ||
-        status.Status === StatusString.STARTED ||
-        status.Status === StatusString.SUBMITTED
-      ) {
-        await sleep(executionConfigs?.waitTimeMs ?? 2000);
-      } else if (status.Status === StatusString.FINISHED) {
-        AnalyticsRepository.log.info({ sql }, 'Command finished');
-        break;
-      } else {
-        AnalyticsRepository.log.error({ error: status.Error }, 'Unknown status');
-        throw new Error(status.Error);
-      }
-    }
+    await this.executeStatement(deleteSql, AnalyticsRepository.log, { waitTimeMs: 10_000 });
+    await this.executeStatement(vacuumSql, AnalyticsRepository.log, { waitTimeMs: 2_000 });
   }
 }
